Extract code from dhis2
