+++ blsqpy/levels.py
@@ -1,3 +1,4 @@
+from functools import partial
 
 class Levels:
     @staticmethod
@@ -11,3 +12,17 @@ class Levels:
     def max_level(df):
         levels = df.path.apply(lambda x: len(x.split('/'))).max()
         return levels - 1
+
+    @staticmethod
+    def add_uid_levels_columns_from_path_column(df):
+      max_level = Levels.max_level(df)
+
+      def to_level(select_level, x):
+            return Levels.to_level_uid(x[0], select_level)
+
+      for level in range(2, max_level - 1):
+            df["uidlevel"+str(level)] = df[['path']
+                                           ].apply(partial(to_level, level), axis=1)
+
+      df.drop(["path"], axis=1, inplace=True)
+      return df
